Move all system and componenent in another folder and rework property
