+++ engine/GPEngine.vcxproj
@@ -216,6 +216,7 @@
     <ClInclude Include="include\Engine\Core\Tools\BranchPrediction.hpp" />
     <ClInclude Include="include\Engine\Core\Tools\ClassUtility.hpp" />
     <ClInclude Include="include\Engine\Core\Tools\Format.hpp" />
+    <ClInclude Include="include\Engine\ECS\System\SystemsManager.hpp" />
     <ClInclude Include="include\Engine\Intermediate\BehaviourComponent.hpp" />
     <ClInclude Include="include\Engine\Intermediate\BehaviourSystem.hpp" />
     <ClInclude Include="include\Engine\Intermediate\Component.hpp" />
