+++ engine/GPEngine.vcxproj.filters
@@ -345,6 +345,9 @@
     <ClInclude Include="include\Engine\Serialization\xml\xmlUtilities.hpp">
       <Filter>include\Serialization\xml</Filter>
     </ClInclude>
+    <ClInclude Include="include\Engine\ECS\System\SystemsManager.hpp">
+      <Filter>include</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include=".clang-format" />
