+++ engine/include/Engine/ECS/Component/Light/PointLight.hpp
@@ -11,7 +11,7 @@
 
 #include "Engine/Core/Tools/ClassUtility.hpp"
 #include "Engine/Intermediate/GameObject.hpp"
-#include "Engine/Resources/Light/Light.hpp"
+#include "Engine/ECS/Component/Light/Light.hpp"
 #include "GPM/Vector3.hpp"
 
 namespace GPE
