+++ engine/include/Engine/ECS/Component/Light/SpotLight.hpp
@@ -10,7 +10,7 @@
 
 #include "Engine/Core/Tools/ClassUtility.hpp"
 #include "Engine/Intermediate/GameObject.hpp"
-#include "Engine/Resources/Light/PointLight.hpp"
+#include "Engine/ECS/Component/Light/PointLight.hpp"
 #include "GPM/Vector3.hpp"
 
 namespace GPE
