+++ engine/include/Engine/Intermediate/GameObject.hpp
@@ -6,11 +6,11 @@
 
 #pragma once
 
-#include "Engine/Intermediate/TransformComponent.hpp" //TransformComponent
-#include <list>                                       //std::list
-#include <memory>                                     //std::unique_ptr
-#include <string>                                     //std::string
-#include <vector>                                     //std::vector
+#include "Engine/ECS/Component/TransformComponent.hpp" //TransformComponent
+#include <list>                                        //std::list
+#include <memory>                                      //std::unique_ptr
+#include <string>                                      //std::string
+#include <vector>                                      //std::vector
 
 // in Inl
 #include "Engine/Intermediate/DataChunk.hpp"
