+++ engine/src/SceneRenderSystem.cpp
@@ -9,13 +9,13 @@
 //#include "Engine/Core/System/TimeSystem.hpp"
 #include "Engine/Core/Rendering/Renderer/RendererGLFW_GL46.hpp"
 #include "Engine/Core/Rendering/Window/WindowGLFW.hpp"
-#include "Engine/Core/System/SystemsManager.hpp"
 #include "Engine/Core/Tools/BranchPrediction.hpp"
+#include "Engine/ECS/Component/Camera.hpp"
+#include "Engine/ECS/Component/Light/Light.hpp"
+#include "Engine/ECS/Component/Model.hpp"
 #include "Engine/ECS/System/RenderSystem.hpp"
-#include "Engine/Resources/Camera.hpp"
-#include "Engine/Resources/Light/Light.hpp"
+#include "Engine/ECS/System/SystemsManager.hpp"
 #include "Engine/Resources/Mesh.hpp"
-#include "Engine/Resources/Model.hpp"
 #include "Engine/Resources/RenderBuffer.hpp"
 #include "Engine/Resources/RenderTexture.hpp"
 #include "Engine/Resources/Shader.hpp"
