+++ engine/src/SpotLight.cpp
@@ -1,4 +1,4 @@
-﻿#include "Engine/Resources/Light/SpotLight.hpp"
+﻿#include "Engine/ECS/Component/Light/SpotLight.hpp"
 #include "GPM/Constants.hpp"
 
 using namespace GPE;
